split up companydashboard class
