+++ spec/cassettes/FinancialMetrics/should_show_burn.yml
