+++ spec/cassettes/FinancialMetrics/should_show_the_correct_kpi_percentage.yml
