+++ spec/cassettes/FinancialMetrics/should_show_the_correct_unlocked_value.yml
